+++ cypress/integration/pages/storyPage/tests.js
@@ -1,6 +1,8 @@
 import pathOr from 'ramda/src/pathOr';
 import path from 'ramda/src/path';
 import getDataUrl from '../../../support/helpers/getDataUrl';
+import topicTagsTest from '../../../support/helpers/topicTagsTest';
+import envConfig from '../../../support/config/envs';
 
 // For testing important features that differ between services, e.g. Timestamps.
 // We recommend using inline conditional logic to limit tests to services which differ.
@@ -45,7 +47,15 @@ export const testsThatFollowSmokeTestConfig = ({ service, pageType }) => {
         }
       });
     });
-
+    it('FOR /news/uk-60508671.amp ONLY - should render topic tags if they are in the json, and they should navigate to correct topic page', () => {
+      if (service === 'news' && Cypress.env('APP_ENV') !== 'local') {
+        const url = '/news/uk-60508671.amp?renderer_env=live';
+        cy.visit(`${envConfig.baseUrl}${url}`);
+        topicTagsTest();
+      } else {
+        cy.log('Test is only for /news/uk-60508671.amp');
+      }
+    });
     it.skip('should render podcast promo if in json and should navigate to correct podcast page', () => {
       cy.log(service);
       if (Cypress.env('APP_ENV') !== 'local') {
